combat: move non-ui state code into its own model object (#152)
* move combat state to its own object

* move model code to its own file

* counter is a property of the screen

* rm
